Isolate GitHub panel logic from UI tests
